+++ tests/testthat/test-cholesky_root.R
@@ -5,4 +5,5 @@ test_that("Cholesky decomposition works", {
   chol <- cov_to_chol(cov)
   expect_true(all.equal(cov, chol_to_cov(chol)))
   expect_equal(unique_chol(numeric()), numeric())
+  expect_equal(cov_to_chol(matrix(0, 2, 2)), c(0, 0, 0))
 })